refactor: 重构脚本逻辑
