+++ src/BootstrapBlazor/wwwroot/modules/utility.js
@@ -674,6 +674,19 @@ export function getHtml(options) {
     return html;
 }
 
+export function getAutoThemeValue() {
+    return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light'
+}
+
+export function setTheme(theme) {
+    if (theme === 'auto' && window.matchMedia('(prefers-color-scheme: dark)').matches) {
+        document.documentElement.setAttribute('data-bs-theme', 'dark')
+    }
+    else {
+        document.documentElement.setAttribute('data-bs-theme', theme);
+    }
+}
+
 export {
     autoAdd,
     autoRemove,
